no more direct access to navGrid cells: accessors return a bool (obstacle), position or size

+++ model/NavCell.hpp
@@ -21,7 +21,7 @@ public:
   bool obstacle;
   uV2 grid_position;
   unsigned int cost;
-  size_t height;
+  float height;
 
 
   //! METHODS
